refactor: refactor utils file structure
