+++ utils/caches/listeners.py
@@ -7,6 +7,7 @@
 # =================================================================================================
 #    Date      Name                    Description of Change
 # 27-May-2022  Wayne Shih              Initial create
+# 30-May-2022  Wayne Shih              Refactor utils file structure
 # $HISTORY$
 # =================================================================================================
 
@@ -14,5 +15,5 @@
 # <Wayne Shih> 26-May-2022
 # https://docs.djangoproject.com/en/3.1/topics/signals/#receiver-functions
 def invalidate_object_cache(sender, instance, **kwargs):
-    from utils.memcached_helpers import MemcachedHelper
+    from utils.caches.memcached_helpers import MemcachedHelper
     MemcachedHelper.invalidate_object_cache(instance.__class__, instance.id)
