+++ utils/caches/redis_client.py
@@ -7,6 +7,7 @@
 # =================================================================================================
 #    Date      Name                    Description of Change
 # 28-May-2022  Wayne Shih              Initial create
+# 30-May-2022  Wayne Shih              Refactor utils file structure
 # $HISTORY$
 # =================================================================================================
